Compare location.pathname >= 0
The String.prototype.indexOf method returns `-1` if the value isn't found as a substring, not `false`. Since `!!-1 === true` in JavaScript, this means that the only condition where the user isn't forced to scroll would be if `location.href.indexOf('grabbag') == 0`, which isn't going to happen unless we're using `grabbag://` or instead of HTTP(S)/FTP/etc.

Secondly, this commits uses `location.pathname` to ignore both the protocol handler and the domain, which is just more surface area for bugs (especially if we're changing `/etc/hosts` or anything for local testing).

Lastly, this commit removes the garbage `console.log` for debugging in production.