let the cowboys go home; fixes issue #26
-Don't overwrite data-src with src if data-src already exists
-Removes data-src after images are reset

Much like the cowboys, this code is a bit on the dangerous side.
If an img has a data-src attribute added anywhere else on the page,
it may never get replaced with a cowboy. If anything else touches
the data-src attribute after the original src is cached, the
cowboy may never leave