Add attribution for Erica Joy quote
Hi Paul,

Congrats on publishing "What Is Code"! All my friends and colleagues have lots of good things to say about it, and it's definitely on my "to read" list.

One of those friends mentioned that you quoted a bit from the Medium post I wrote about being a black woman in tech (if I'm being completely transparent, it was message in Slack with a link to your piece and "CTRL+F medium").

I appreciate that you took the time to read my post and felt compelled to share a bit of it. It took a lot out of me both to live my experiences in tech and to recall them and share them with the world. For that reason, it was and remains important to me that I not be erased from my experience, so I licensed my post CC-BY-ND so that people could share my story with attribution. I'm respectfully requesting that you honor that by attributing my quote to me in your piece.

Thanks so much for your understanding and again for sharing my story. Looking forward to reading your entire piece. Have a great weekend!

Erica Joy Baker
@ericajoy
http://www.ericabaker.com