Women aren't unintelligent
Before this change, the end of this paragraph made it seem like you
think women are unintelligent - the implied meaning was that *if*
programming actually *was* the most intellectually demanding task
imaginable, *then* it would be reasonable to assume women couldn't
be as good as men.
I believe (hope) that this is really not what you
meant; this change makes that much more clear.