Remove the controversial sentence alltogether
After coming back to read this again, I realized this is probably fixed better
by just removing the offending sentence alltogether. I don't think any value
is lost this way, but there is now no way to read this paragraph as if it means
that the author claims women are stupid.