Fixes #125. (outerHTML is a thing!!)
Lots could be done with thing but I'll leave it for now. Like showing
the whole article, or letting you choose what to display, or... some
kind of vertical fisheye... man, remember when I had the whole thing
fisheyeing? uh it would be cool to have shortcut buttons to traverse
the tree, expanding/collapsing to a certain spot... and highlighting
what page thing it corresponded to... but.....